Change `floorSize` data type to `Float`

+++ app/src/main/java/com/anafthdev/snakeclassic/model/GameConfigurationData.kt
@@ -1,7 +1,7 @@
 package com.anafthdev.snakeclassic.model
 
 data class GameConfigurationData(
-	val floorSize: Int,
+	val floorSize: Float,
 	val movementDelay: Int,
 	val easingAnimationDelay: Int
 )
